move: useJoin 함수를 useAuth로 이동
