+++ app/hooks/queries/useAuth.ts
@@ -161,4 +161,34 @@ const useSocialLogin = () => {
     },
   });
 };
-export { useLogin, useKakaoLogin, useKakaoProfile, useSocialLogin };
+
+interface JoinProps {
+  memberId: string;
+  password: string;
+  name: string;
+  email: string;
+}
+
+const useJoin = () => {
+  const router = useRouter();
+  return useMutation({
+    mutationFn: async ({ memberId, password, name, email }: JoinProps) => {
+      const response: Response<JoinProps> = await Instance.post('/signup', {
+        memberId,
+        password,
+        name,
+        email,
+      });
+      return response;
+    },
+    onSuccess: (data) => {
+      router.push('/login');
+    },
+    onError: (error) => {
+      console.error('useJoin: ', error.message);
+      throw error;
+    },
+  });
+};
+
+export { useLogin, useKakaoLogin, useKakaoProfile, useSocialLogin, useJoin };
